ansible/kvm: disable the singlestep test
It seems that the nested virtualization setup on Travis runs an Intel
guest on top of a AMD guest. KVMi might report features (eg. singlestep)
which are not available for AMD processors.

Setting
        libvirt.cpu_model = "host-passthrough"
didn't change anything.

This is a workaround until we find a way to avoid reporting such features.

Signed-off-by: Adalbert Lazăr <alazar@bitdefender.com>